requirements.txt
apache-beam[gcp]
black
google-cloud-bigquery
pandas
pyarrow
pylint
pytest
pytest-console-scripts